Sem 8/Computer Vision/requirements.txt
pillow
numpy
opencv-python

# dev
black 
isort